linux: Cast usbfs_iso_packet_desc status before comparing
Whereas the URB status field is signed int, the Linux USBDEVFS API
declares the status field of the iso packet descriptor as unsigned
int. We compare them with signed values in the switch cases, because the
error values are negative (system error macros like ENOENT are
positive).

https://docs.kernel.org/driver-api/usb/usb.html
https://docs.kernel.org/driver-api/usb/error-codes.html

This avoids a -Wsign-conversion warning.

+++ libusb/os/linux_usbfs.c
@@ -2518,7 +2518,7 @@ static int handle_iso_completion(struct usbi_transfer *itransfer,
 			&transfer->iso_packet_desc[tpriv->iso_packet_offset++];
 
 		lib_desc->status = LIBUSB_TRANSFER_COMPLETED;
-		switch (urb_desc->status) {
+		switch ((int)urb_desc->status) {
 		case 0:
 			break;
 		case -ENOENT: /* cancelled */
